browserND/requirements.txt
alembic==1.4.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
cycler==0.10.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-SSE==1.0.0
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
opencv-contrib-python==4.5.1.48
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
redis==3.5.3
requests==2.24.0
setproctitle==1.2.1
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.10
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
